pyproject.toml
[project]
name = "getall-ai"
version = "0.1.3.post6"
description = "GetAll - AI-native crypto trading pet agent"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "GetAll contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # --- CLI & display ---
    "typer>=0.9.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    # --- LLM ---
    "litellm>=1.0.0",
    # --- Config ---
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # --- HTTP API ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # --- Cache / queue / locks ---
    "redis[hiredis]>=5.0.0",
    # --- Trading extensions ---
    "ccxt>=4.0",
    "pandas>=2.0.0",
    "pandas-ta>=0.4.67b0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    # --- Networking ---
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx[socks]>=0.25.0",
    # --- Crypto ---
    "cryptography>=42.0.0",
    # --- Logging ---
    "loguru>=0.7.0",
    # --- Web tools ---
    "readability-lxml>=0.8.0",
    # --- Scheduling ---
    "croniter>=2.0.0",
    # --- Chat channels ---
    "dingtalk-stream>=0.4.0",
    "python-telegram-bot[socks]>=21.0",
    "lark-oapi>=1.0.0",
    "socksio>=1.0.0",
    "python-socketio>=5.11.0",
    "msgpack>=1.0.8",
    "slack-sdk>=3.26.0",
    "qq-botpy>=1.0.0",
    "python-socks[asyncio]>=2.4.0",
]

[project.optional-dependencies]
backtest = [
    "nautilus_trader>=1.200.0,<1.209.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
getall = "getall.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["getall"]

[tool.hatch.build.targets.wheel.sources]
"getall" = "getall"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "getall/**/*.py",
    "getall/skills/**/*.md",
    "getall/skills/**/*.sh",
    "getall/trading/skills/**/*.md",
    "getall/trading/workspace_templates/**/*.md",
    "getall/trading/cron_seeds.yaml",
]

[tool.hatch.build.targets.sdist]
include = [
    "getall/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "getall/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
